riskoddscalc expected value calculator
Haven't changed the site on Github Pages, but I've created a script to find the expected losses (and the probabilities of winning as well). Thought of a new algorithm, implemented it in Fortran because I wanted to learn the language and figured that this would be a way to do it. You'll have to compile it yourself.

Also included here is a python script that uses a Monte Carlo simulation to calculate the same thing. Just wanted to check my work.